+++ autol_driver/src/packet_parser/g32_v2_parser.hpp
@@ -180,7 +180,7 @@ void G32V2Parser::ChangeFovToPcd(std::vector<AutoLG32V2FovDataBlock> &fov_data_s
 {
     const int32_t numOfChannel = 16;
     float intensity = 0;
-    double timestamp = 0;
+    uint64_t timestamp = 0;
 
     for (int32_t i = 0; i < (int32_t)fov_data_set_t.size(); i++)
     {
